requirements.txt
streamlit
requests
langchain
langchain-google-genai
pypdf
chromadb
sentence-transformers
pandas
python-dotenv
google-generativeai
langchain_community